[Server] / #80 / add resource auth

+++ server/src/routes/routes.service.ts
@@ -4,6 +4,7 @@ import {
   Injectable,
   InternalServerErrorException,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { QueryFailedError, Repository, UpdateResult } from 'typeorm';
@@ -22,6 +23,9 @@ import fs from 'fs';
 import { join } from 'path';
 import { plainToClass } from 'class-transformer';
 import { validate } from 'class-validator';
+import { UsersService } from 'src/users/users.service';
+import jwt, { JsonWebTokenError } from 'jsonwebtoken';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class RoutesService {
@@ -33,10 +37,16 @@ export class RoutesService {
     private pinsRepository: Repository<PinEntity>,
     @InjectRepository(PictureEntity)
     private picturesRepository: Repository<PictureEntity>,
+    private readonly usersService: UsersService,
+    private configService: ConfigService,
   ) {}
 
-  async getUserRoutes(page: number): Promise<object> {
-    //TODO where문으로 해당 유저의 루트만 가져오기
+  async getUserRoutes(page: number, accessToken: string | undefined) {
+    const decode = jwt.verify(
+      accessToken,
+      this.configService.get<string>('ACCESS_SECRET'),
+    );
+
     //루트의 createdAt, 핀의 ranking컬럼, 사진의 아이디 순으로 오름차순 정렬
     const routes = await this.routesRepository
       .createQueryBuilder('Routes') //alias -> SELECT Routes as Routes
@@ -57,6 +67,7 @@ export class RoutesService {
         'Pins.locationName',
         'Pictures.fileName',
       ])
+      .where('Routes.userId = :userId', { userId: decode['id'] })
       .orderBy('Routes.createdAt')
       .addOrderBy('Pins.ranking')
       .addOrderBy('Pictures.id')
@@ -87,8 +98,17 @@ export class RoutesService {
   }
 
   // 핀을 추가하기 위해서는 새로 생성된 루트의 아이디가 필요하므로, 루트를 생성하고, 루트 아이디를 이용해 핀들을 생성한다.
-  async createRoute(routeStr: string, files: Array<Express.Multer.File>) {
+  async createRoute(
+    routeStr: string,
+    files: Array<Express.Multer.File>,
+    accessToken: string | undefined,
+  ) {
     try {
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
+
       //문자열 JSON을 parse한 뒤, PatchPinDto타입의 객체를 생성한다.
       const routePins = plainToClass(PostRouteDto, JSON.parse(routeStr));
 
@@ -120,12 +140,11 @@ export class RoutesService {
         }
       }
 
-      //TODO 유저 아이디 바꾸기
       //public은 예약어이다
       const { routeName, description, color, time } = routePins;
       const { pins } = routePins;
       const newRoute = await this.routesRepository.save({
-        userId: 1,
+        userId: decode['id'],
         routeName: routeName,
         description: description,
         public: routePins.public,
@@ -158,7 +177,6 @@ export class RoutesService {
       pins.forEach((pin) => {
         pin['routesId'] = newRoute.id;
       });
-      // console.log(pins);
 
       //새 핀들 생성
       //bulk insert
@@ -188,37 +206,55 @@ export class RoutesService {
         throw new NotFoundException();
       } else if (err.status === 400 || err instanceof SyntaxError) {
         throw new BadRequestException(null, err.message);
+      } else if (err instanceof JsonWebTokenError) {
+        throw err;
       } else {
         throw new InternalServerErrorException();
       }
     }
   }
-  //TODO: 해당 사용자가 작성한 루트만 수정할 수 있어야 한다.
-  async updateRoute(routeId: number, route: PatchRouteDto) {
+
+  async updateRoute(
+    routeId: number,
+    route: PatchRouteDto,
+    accessToken: string | undefined,
+  ) {
     try {
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
       //루트 아이디와 일치하는 요소 업데이트
       const result = await this.routesRepository
         .createQueryBuilder()
         .update('Routes')
         .set(route)
-        .where('id = :id', { id: routeId })
+        .where('id = :id AND userId = :userId', {
+          id: routeId,
+          userId: decode['id'],
+        })
         .execute();
 
       if (!result.affected) {
         //없는 루트, 또는 다른 유저가 작성한 루트를 업데이트 하려는 경우
-        throw new NotFoundException();
+        throw new UnauthorizedException();
       }
     } catch (err) {
-      if (err.status === 404) {
-        throw new NotFoundException();
+      if (err.status === 401) {
+        throw err;
       } else {
         throw new InternalServerErrorException();
       }
     }
   }
 
-  async deleteRoute(routeId: number) {
+  async deleteRoute(routeId: number, accessToken: string | undefined) {
     try {
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
+
       //루트를 지우기 전, 루트에 속한 사진들의 정보를 가져온다.
       const picturesInfo = await this.picturesRepository
         .createQueryBuilder('Pictures')
@@ -233,12 +269,15 @@ export class RoutesService {
         .createQueryBuilder()
         .delete()
         .from('Routes')
-        .where('id = :id', { id: routeId })
+        .where('id = :id AND userId = :userId', {
+          id: routeId,
+          userId: decode['id'],
+        })
         .execute();
 
       if (!result.affected) {
         //없는 루트, 또는 다른 유저가 작성한 루트를 삭제 하려는 경우
-        throw new NotFoundException();
+        throw new UnauthorizedException();
       }
 
       //사진 파일 삭제
@@ -257,8 +296,8 @@ export class RoutesService {
       const updatePinInfoId = redefineTooClose(dbPins);
       await this.pinsRepository.save(updatePinInfoId);
     } catch (err) {
-      if (err.status === 404) {
-        throw new NotFoundException();
+      if (err.status === 401) {
+        throw err;
       } else {
         //테이블의 정보는 삭제 되더라도 실제 파일이 없는 경우 이 에러가 발생한다. (테이블에서는 삭제 처리된다.)
         throw new InternalServerErrorException();
@@ -266,39 +305,82 @@ export class RoutesService {
     }
   }
 
-  async getPins(routeId: number): Promise<PinEntity[]> {
-    //TODO루트가 해당 사용자 소유인지 확인하기
+  async getPins(
+    routeId: number,
+    accessToken: string | undefined,
+  ): Promise<PinEntity[]> {
+    try {
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
 
-    const pins = await this.pinsRepository
-      .createQueryBuilder('Pins') //alias -> SELECT Pins as Pins
-      .leftJoinAndSelect('Pins.Pictures', 'Pictures')
-      .where('Pins.routesId = :id', { id: routeId })
-      .orderBy('Pins.ranking')
-      .addOrderBy('Pictures.id')
-      .getMany(); //여러 개의 결과를 가져온다. entity를 반환한다. getRawMany등으로 raw data를 가져올 수 있다.
+      //빈 배열이 나오는 경우 -> 없는 루트를 조회하거나, 자기가 저장하지 않은 루트를 조회 또는 핀이 저장되지 않은 루트를 저장하는 경우 -> 응답방식 생각해 보기
+      const pins = await this.pinsRepository
+        .createQueryBuilder('Pins') //alias -> SELECT Pins as Pins
+        .leftJoinAndSelect('Pins.Pictures', 'Pictures')
+        .leftJoin('Pins.Routes', 'Routes')
+        .where('Pins.routesId = :id AND Routes.userId = :userId', {
+          id: routeId,
+          userId: decode['id'],
+        })
+        .orderBy('Pins.ranking')
+        .addOrderBy('Pictures.id')
+        .getMany(); //여러 개의 결과를 가져온다. entity를 반환한다. getRawMany등으로 raw data를 가져올 수 있다.
 
-    return pins;
+      return pins;
+    } catch (err) {
+      throw err;
+    }
   }
 
   async updatePin(
     routeId: number,
     pinId: number,
-    pin: PatchPinDto,
+    pinStr: string,
     files: Array<Express.Multer.File>,
-  ): Promise<UpdateResult> {
+    accessToken: string | undefined,
+  ) {
     try {
+      //문자열 JSON을 parse한 뒤, PatchPinDto타입의 객체를 생성한다.
+      const pin = plainToClass(PatchPinDto, JSON.parse(pinStr));
+
+      //새로 생성한 객체의 유효성 검증
+      //유효하지 않은 키가 있으면 pinValError 배열에 추가된다.
+      const pinValError = await validate(pin, {
+        forbidUnknownValues: true,
+      });
+      if (pinValError.length > 0) {
+        const fstVal =
+          pinValError[0].constraints[
+            Object.keys(pinValError[0].constraints)[0]
+          ];
+        throw new BadRequestException(null, fstVal);
+      }
+
+      //토큰의 유효성 검사
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
+
       const result = await this.pinsRepository
-        .createQueryBuilder()
+        .createQueryBuilder('Pins')
+        .leftJoin('Pins.Routes', 'Routes')
         .update('Pins')
         .set(pin)
-        .where('routesId = :routesId AND id = :id', {
-          routesId: routeId,
-          id: pinId,
-        })
+        .where(
+          'Pins.routesId = :routesId AND Pins.id = :id AND Routes.userId = :userId',
+          {
+            routesId: routeId,
+            id: pinId,
+            userId: decode['id'],
+          },
+        )
         .execute();
 
       if (!result.affected) {
-        throw new NotFoundException();
+        throw new UnauthorizedException();
       }
 
       const newPictures = [];
@@ -328,18 +410,29 @@ export class RoutesService {
       //1. select문으로 updatePinInfoId의 핀들을 조회한다
       //2. 변경 사항이 있는 레코드에만 update쿼리가 실행된다.(변경 사항이 없으면 update쿼리는 실행되지 않는다.)
       await this.pinsRepository.save(updatePinInfoId);
-      return result;
     } catch (err) {
-      if (err.status === 404) {
-        throw new NotFoundException();
+      if (err.status === 401) {
+        throw err;
+      } else if (err.status === 400 || err instanceof SyntaxError) {
+        //잘못된 JSON 형식을 받을 경우
+        throw new BadRequestException(null, err.message);
       } else {
         throw new InternalServerErrorException();
       }
     }
   }
 
-  async deletePin(routeId: number, pinId: number) {
+  async deletePin(
+    routeId: number,
+    pinId: number,
+    accessToken: string | undefined,
+  ) {
     try {
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
+
       //핀을 지우기 전, 핀에 속한 사진들의 정보를 가져온다.
       const picturesInfo = await this.picturesRepository
         .createQueryBuilder('Pictures')
@@ -347,18 +440,26 @@ export class RoutesService {
         .where('pinId = :pinId', { pinId: pinId })
         .getMany();
 
+      //삭제된 핀의 정보
       const result = await this.pinsRepository
-        .createQueryBuilder()
-        .delete()
-        .from('Pins')
-        .where('routesId = :routesId AND id = :id', {
-          routesId: routeId,
-          id: pinId,
-        })
-        .execute();
+        .createQueryBuilder('Pins')
+        .innerJoin('Pins.Routes', 'Routes')
+        .where(
+          'Pins.routesId = :routesId AND Pins.id = :id AND Routes.userId = :userId',
+          {
+            routesId: routeId,
+            id: pinId,
+            userId: decode['id'],
+          },
+        )
+        .getMany()
+        .then((Pins) => {
+          return this.pinsRepository.remove(Pins);
+        });
+
       //없는 핀, 또는 다른 유저가 작성한 핀을 삭제 하려는 경우
-      if (!result.affected) {
-        throw new NotFoundException();
+      if (!result.length) {
+        throw new UnauthorizedException();
       }
 
       //사진 파일 삭제
@@ -379,8 +480,10 @@ export class RoutesService {
 
       return result;
     } catch (err) {
-      if (err.status === 404) {
-        throw new NotFoundException();
+      if (err.status === 401) {
+        throw err;
+      } else if (err instanceof JsonWebTokenError) {
+        throw err;
       } else {
         //테이블의 정보는 삭제 되더라도 실제 파일이 없는 경우 이 에러가 발생한다. (테이블에서는 삭제 처리된다.)
         throw new InternalServerErrorException();
@@ -390,11 +493,45 @@ export class RoutesService {
 
   async createPin(
     routeId: number,
-    pin: PatchPinDto,
+    pinStr: string,
     files: Array<Express.Multer.File>,
+    accessToken: string | undefined,
   ) {
     try {
+      //문자열 JSON을 parse한 뒤, PatchPinDto타입의 객체를 생성한다.
+      const pin = plainToClass(PatchPinDto, JSON.parse(pinStr));
+
+      //새로 생성한 객체의 유효성 검증
+      //유효하지 않은 키가 있으면 pinValError 배열에 추가된다.
+      const pinValError = await validate(pin, {
+        forbidUnknownValues: true,
+      });
+      if (pinValError.length > 0) {
+        const fstVal =
+          pinValError[0].constraints[
+            Object.keys(pinValError[0].constraints)[0]
+          ];
+        throw new BadRequestException(null, fstVal);
+      }
+
+      const decode = jwt.verify(
+        accessToken,
+        this.configService.get<string>('ACCESS_SECRET'),
+      );
+
+      //핀을 추가하려는 루트가 해당 사용자 소유인지 확인한다.
+      const selectRouteResult = await this.routesRepository.findOne({
+        userId: decode['id'],
+        id: routeId,
+      });
+      if (!selectRouteResult) {
+        throw new UnauthorizedException();
+      }
+
+      //핀을 생성한다.
       const newPin = { ...pin, routesId: routeId };
+      const createPinResult = await this.pinsRepository.save(newPin);
+
       const dbPins = await this.pinsRepository
         .createQueryBuilder('Pins')
         .select(['Pins.id', 'Pins.latitude', 'Pins.longitude', 'Pins.tooClose'])
@@ -408,7 +545,6 @@ export class RoutesService {
       await this.pinsRepository.save(dbPinId);
 
       //생성한 핀을 이용해 사진의 정보를 저장한다.
-      const createPinResult = await this.pinsRepository.save(newPin);
       const newPictures = [];
       files.forEach((file) => {
         newPictures.push({
@@ -422,8 +558,14 @@ export class RoutesService {
       //잘못된 칼럼 정보를 추가하려는 경우
       if (err instanceof QueryFailedError) {
         throw new BadRequestException();
+      } else if (err.status === 400 || err instanceof SyntaxError) {
+        //잘못된 JSON 형식을 받을 경우
+        throw new BadRequestException(null, err.message);
+      } else if (err.status === 401) {
+        throw err;
+      } else {
+        throw new InternalServerErrorException();
       }
-      throw new InternalServerErrorException();
     }
   }
 }
